requirements.txt
aiofiles==0.7.0
aiohttp==3.8.1
binance-python==1.0.16
certifi==2021.10.8
numpy==1.22.3
pandas==1.4.2
psutil==5.9.0
python-binance==1.0.16
python-dotenv==0.20.0
pyyaml==6.0
rich==12.0.0
ta==0.9.0